+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/test/java/mozilla/components/browser/engine/gecko/prompt/GeckoPromptDelegateTest.kt
@@ -96,7 +96,7 @@ class GeckoPromptDelegateTest {
 
         assertTrue(promptRequestSingleChoice is SingleChoice)
 
-        (promptRequestSingleChoice as SingleChoice).onSelect(mock())
+        (promptRequestSingleChoice as SingleChoice).onConfirm(mock())
         assertTrue(confirmWasCalled)
     }
 
@@ -141,7 +141,7 @@ class GeckoPromptDelegateTest {
 
         assertTrue(promptRequestSingleChoice is MultipleChoice)
 
-        (promptRequestSingleChoice as MultipleChoice).onSelect(arrayOf())
+        (promptRequestSingleChoice as MultipleChoice).onConfirm(arrayOf())
         assertTrue(confirmWasCalled)
     }
 
@@ -183,7 +183,7 @@ class GeckoPromptDelegateTest {
 
         assertTrue(promptRequestSingleChoice is PromptRequest.MenuChoice)
 
-        (promptRequestSingleChoice as PromptRequest.MenuChoice).onSelect(mock())
+        (promptRequestSingleChoice as PromptRequest.MenuChoice).onConfirm(mock())
         assertTrue(confirmWasCalled)
     }
 
@@ -224,7 +224,7 @@ class GeckoPromptDelegateTest {
         (alertRequest as PromptRequest.Alert).onDismiss()
         assertTrue(dismissWasCalled)
 
-        (alertRequest as PromptRequest.Alert).onShouldShowNoMoreDialogs(true)
+        (alertRequest as PromptRequest.Alert).onConfirm(true)
         assertTrue(setCheckboxValueWasCalled)
 
         assertEquals((alertRequest as PromptRequest.Alert).title, "title")
@@ -286,7 +286,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_DATE, "", "", "", callback)
         assertTrue(dateRequest is PromptRequest.TimeSelection)
-        (dateRequest as PromptRequest.TimeSelection).onSelect(Date())
+        (dateRequest as PromptRequest.TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as PromptRequest.TimeSelection).title, "title")
 
@@ -331,7 +331,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2019-11-30".toDate("yyyy-MM-dd"))
         }
         val selectedDate = "2019-11-28".toDate("yyyy-MM-dd")
-        (timeSelectionRequest as PromptRequest.TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as PromptRequest.TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-MM-dd")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -359,7 +359,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_MONTH, "", "", "", callback)
         assertTrue(dateRequest is PromptRequest.TimeSelection)
-        (dateRequest as PromptRequest.TimeSelection).onSelect(Date())
+        (dateRequest as PromptRequest.TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -401,7 +401,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2019-11".toDate("yyyy-MM"))
         }
         val selectedDate = "2019-11".toDate("yyyy-MM")
-        (timeSelectionRequest as PromptRequest.TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as PromptRequest.TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-MM")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -429,7 +429,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_WEEK, "", "", "", callback)
         assertTrue(dateRequest is PromptRequest.TimeSelection)
-        (dateRequest as PromptRequest.TimeSelection).onSelect(Date())
+        (dateRequest as PromptRequest.TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -471,7 +471,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2018-W26".toDate("yyyy-'W'ww"))
         }
         val selectedDate = "2018-W26".toDate("yyyy-'W'ww")
-        (timeSelectionRequest as PromptRequest.TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as PromptRequest.TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-'W'ww")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -500,7 +500,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_TIME, "", "", "", callback)
         assertTrue(dateRequest is PromptRequest.TimeSelection)
-        (dateRequest as PromptRequest.TimeSelection).onSelect(Date())
+        (dateRequest as PromptRequest.TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -542,7 +542,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "18:00".toDate("HH:mm"))
         }
         val selectedDate = "17:00".toDate("HH:mm")
-        (timeSelectionRequest as PromptRequest.TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as PromptRequest.TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("HH:mm")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -574,7 +574,7 @@ class GeckoPromptDelegateTest {
             GeckoSession.PromptDelegate.DATETIME_TYPE_DATETIME_LOCAL, "", "", "", callback
         )
         assertTrue(dateRequest is PromptRequest.TimeSelection)
-        (dateRequest as PromptRequest.TimeSelection).onSelect(Date())
+        (dateRequest as PromptRequest.TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as PromptRequest.TimeSelection).title, "title")
     }
@@ -616,7 +616,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2018-06-14T00:00".toDate("yyyy-MM-dd'T'HH:mm"))
         }
         val selectedDate = "2018-06-12T19:30".toDate("yyyy-MM-dd'T'HH:mm")
-        (timeSelectionRequest as PromptRequest.TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as PromptRequest.TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-MM-dd'T'HH:mm")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as PromptRequest.TimeSelection).title, "title")
     }