+++ mobile/android/android-components/components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/prompt/GeckoPromptDelegateTest.kt
@@ -95,7 +95,7 @@ class GeckoPromptDelegateTest {
 
         assertTrue(promptRequestSingleChoice is SingleChoice)
 
-        (promptRequestSingleChoice as SingleChoice).onSelect(mock())
+        (promptRequestSingleChoice as SingleChoice).onConfirm(mock())
         assertTrue(confirmWasCalled)
     }
 
@@ -140,7 +140,7 @@ class GeckoPromptDelegateTest {
 
         assertTrue(promptRequestSingleChoice is MultipleChoice)
 
-        (promptRequestSingleChoice as MultipleChoice).onSelect(arrayOf())
+        (promptRequestSingleChoice as MultipleChoice).onConfirm(arrayOf())
         assertTrue(confirmWasCalled)
     }
 
@@ -180,7 +180,7 @@ class GeckoPromptDelegateTest {
 
         assertTrue(promptRequestSingleChoice is PromptRequest.MenuChoice)
 
-        (promptRequestSingleChoice as PromptRequest.MenuChoice).onSelect(mock())
+        (promptRequestSingleChoice as PromptRequest.MenuChoice).onConfirm(mock())
         assertTrue(confirmWasCalled)
     }
 
@@ -221,7 +221,7 @@ class GeckoPromptDelegateTest {
         (alertRequest as PromptRequest.Alert).onDismiss()
         assertTrue(dismissWasCalled)
 
-        (alertRequest as PromptRequest.Alert).onShouldShowNoMoreDialogs(true)
+        (alertRequest as PromptRequest.Alert).onConfirm(true)
         assertTrue(setCheckboxValueWasCalled)
 
         assertEquals((alertRequest as PromptRequest.Alert).title, "title")
@@ -283,7 +283,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_DATE, "", "", "", callback)
         assertTrue(dateRequest is TimeSelection)
-        (dateRequest as TimeSelection).onSelect(Date())
+        (dateRequest as TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as TimeSelection).title, "title")
 
@@ -328,7 +328,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2019-11-30".toDate("yyyy-MM-dd"))
         }
         val selectedDate = "2019-11-28".toDate("yyyy-MM-dd")
-        (timeSelectionRequest as TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-MM-dd")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as TimeSelection).title, "title")
     }
@@ -356,7 +356,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_MONTH, "", "", "", callback)
         assertTrue(dateRequest is TimeSelection)
-        (dateRequest as TimeSelection).onSelect(Date())
+        (dateRequest as TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as TimeSelection).title, "title")
     }
@@ -398,7 +398,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2019-11".toDate("yyyy-MM"))
         }
         val selectedDate = "2019-11".toDate("yyyy-MM")
-        (timeSelectionRequest as TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-MM")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as TimeSelection).title, "title")
     }
@@ -426,7 +426,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_WEEK, "", "", "", callback)
         assertTrue(dateRequest is TimeSelection)
-        (dateRequest as TimeSelection).onSelect(Date())
+        (dateRequest as TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as TimeSelection).title, "title")
     }
@@ -468,7 +468,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2018-W26".toDate("yyyy-'W'ww"))
         }
         val selectedDate = "2018-W26".toDate("yyyy-'W'ww")
-        (timeSelectionRequest as TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-'W'ww")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as TimeSelection).title, "title")
     }
@@ -497,7 +497,7 @@ class GeckoPromptDelegateTest {
         })
         promptDelegate.onDateTimePrompt(null, "title", DATETIME_TYPE_TIME, "", "", "", callback)
         assertTrue(dateRequest is TimeSelection)
-        (dateRequest as TimeSelection).onSelect(Date())
+        (dateRequest as TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as TimeSelection).title, "title")
     }
@@ -539,7 +539,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "18:00".toDate("HH:mm"))
         }
         val selectedDate = "17:00".toDate("HH:mm")
-        (timeSelectionRequest as TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("HH:mm")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as TimeSelection).title, "title")
     }
@@ -571,7 +571,7 @@ class GeckoPromptDelegateTest {
             GeckoSession.PromptDelegate.DATETIME_TYPE_DATETIME_LOCAL, "", "", "", callback
         )
         assertTrue(dateRequest is TimeSelection)
-        (dateRequest as TimeSelection).onSelect(Date())
+        (dateRequest as TimeSelection).onConfirm(Date())
         assertTrue(confirmCalled)
         assertEquals((dateRequest as TimeSelection).title, "title")
     }
@@ -613,7 +613,7 @@ class GeckoPromptDelegateTest {
             assertEquals(maximumDate, "2018-06-14T00:00".toDate("yyyy-MM-dd'T'HH:mm"))
         }
         val selectedDate = "2018-06-12T19:30".toDate("yyyy-MM-dd'T'HH:mm")
-        (timeSelectionRequest as TimeSelection).onSelect(selectedDate)
+        (timeSelectionRequest as TimeSelection).onConfirm(selectedDate)
         assertNotNull(geckoDate?.toDate("yyyy-MM-dd'T'HH:mm")?.equals(selectedDate))
         assertEquals((timeSelectionRequest as TimeSelection).title, "title")
     }